Fix org-mind-map-narrow-to-heading-content
I suspect this might have been a typo. The function was returning the
point of the argument "b", rather than the start point of the first
non-matching headline. The result of this was that when
org-mind-map-include-text was enabled, text would only be displayed on
leaf nodes that had no child. For any node with a child, the text
would be skipped.

Fix the issue and update the related screenshots in the README.